Move grading config to judgels.grading package; order below rabbitmq
- Revert v2-ubuntu-* ansible grader changes (only v3-ubuntu-24.04 kept)
- Move JudgelsServerGradingConfiguration and JudgelsGraderGradingConfiguration
  to package judgels.grading
- Order the grading getter/YAML block after rabbitmq in both apps
