+++ deployment/v3-ubuntu-24.04/ansible/conf/judgels-grader.yml.j2
@@ -17,14 +17,14 @@ logging:
 judgels:
   baseDataDir: var/data
 
-  grading:
-    gradingRequestQueueName: gabriel-grading-request
-    numWorkerThreads: {{ gabriel_grading_numWorkerThreads }}
-
   rabbitmq:
     host: {{ groups['core'][0] }}
     username: {{ rabbitmq_username }}
     password: {{ rabbitmq_password }}
+
+  grading:
+    gradingRequestQueueName: gabriel-grading-request
+    numWorkerThreads: {{ gabriel_grading_numWorkerThreads }}
 
 gabriel:
   cache:
