+++ deployment/v3-ubuntu-24.04/ansible/conf/judgels-server.yml.j2
@@ -53,16 +53,16 @@ judgels:
 {% endif %}
     name: {{ app_name }}
 
-  grading:
-    gradingRequestQueueName: judgels-grading-request
-    gradingResponseQueueName: judgels-grading-response
-
 {% if rabbitmq_username is defined %}
   rabbitmq:
     host: localhost
     username: {{ rabbitmq_username }}
     password: {{ rabbitmq_password }}
 {% endif %}
+
+  grading:
+    gradingRequestQueueName: judgels-grading-request
+    gradingResponseQueueName: judgels-grading-response
 
 jophiel:
   auth:
