+++ judgels-backends/judgels-grader-app/src/main/java/judgels/gabriel/grading/GradingModule.java
@@ -9,7 +9,7 @@ import jakarta.inject.Singleton;
 import java.nio.file.Path;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ThreadPoolExecutor;
-import judgels.JudgelsGraderGradingConfiguration;
+import judgels.grading.JudgelsGraderGradingConfiguration;
 import judgels.messaging.MessageListener;
 import judgels.service.JudgelsBaseDataDir;
