+++ judgels-backends/judgels-grader-app/src/main/java/judgels/gabriel/grading/GradingWorker.java
@@ -14,7 +14,6 @@ import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.Map;
 import java.util.Optional;
-import judgels.JudgelsGraderGradingConfiguration;
 import judgels.JudgelsObjectMappers;
 import judgels.gabriel.api.GradingConfig;
 import judgels.gabriel.api.GradingEngine;
@@ -35,6 +34,7 @@ import judgels.gabriel.engines.GradingEngineRegistry;
 import judgels.gabriel.languages.GradingLanguageRegistry;
 import judgels.gabriel.sandboxes.fake.FakeSandboxFactory;
 import judgels.gabriel.sandboxes.isolate.IsolateSandboxFactory;
+import judgels.grading.JudgelsGraderGradingConfiguration;
 import judgels.messaging.MessageClient;
 import judgels.messaging.api.Message;
 import org.slf4j.Logger;
