+++ judgels-backends/judgels-grader-app/var/conf/judgels-grader.yml.example
@@ -12,14 +12,14 @@ logging:
 judgels:
   baseDataDir: var/data
 
-  grading:
-    gradingRequestQueueName: gabriel-grading-request
-    numWorkerThreads: 2
-
   rabbitmq:
     host: localhost
     username: guest
     password: guest
+
+  grading:
+    gradingRequestQueueName: gabriel-grading-request
+    numWorkerThreads: 2
 
 gabriel:
   cache:
