+++ judgels-backends/judgels-server-app/src/main/java/judgels/grading/JudgelsServerGradingConfiguration.java
@@ -1,4 +1,4 @@
-package judgels;
+package judgels.grading;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import org.immutables.value.Value;
